Change the types used in most part of the code to better reflect the values they can hold.

+++ src/color.cpp
@@ -1,6 +1,6 @@
 #include "color.h"
 
-void HSBToRGB( unsigned int i_hue, unsigned int i_saturation, unsigned int i_brightness, unsigned int *o_red, unsigned int *o_green, unsigned int *o_blue )
+void HSBToRGB( int16_t i_hue, uint8_t i_saturation, uint8_t i_brightness, uint8_t *o_red, uint8_t *o_green, uint8_t *o_blue )
 {
   if( i_saturation == 0 )
   {
